# configs/swagger.yaml
openapi: 3.0.0
info:
  title: NestJS - Open API 3.0.0
  version: 2.0.0
servers:
  - url: http://localhost:3000
...
paths:
...
  /locations:
    get:
      tags:
        - Home
      summary: locations
      description: Get all data in HomePage.
      responses:
        200:
          description: Successful refresh
        401:
          description: Access token is missing or invalid
        400:
          description: Invalid Credentials
        500:
          description: Interval Server
...
  /news:
    get:
      tags:
        - Home
      summary: read news
      description: read info Homepage.
      parameters:
        - name: page
          in: query
          description: ID of Share to return
          schema:
            type: string
        - name: limit
          in: query
          description: ID of Share to return
          schema:
            type: string
        - name: sort
          in: query
          description: ID of Share to return
          schema:
            type: string
        - name: search
          in: query
          description: ID of Share to return
          schema:
            type: string
      responses:
        200:
          description: Get all Sharing success
        400:
          description: Invalid status value
  /news/{id}:
    get:
      tags:
        - Home
      summary: read new
      description: read info Homepage.
      parameters:
        - name: id
          in: path
          description: ID of Share to return
          required: true
          schema:
            type: string
      responses:
        200:
          description: Get all Sharing success
        400:
          description: Invalid status value
...
  /admin/locations/{children}:
    post:
      tags:
        - Location
      summary: update info in Location
      parameters:
        - name: children
          description: address | email | phone
          in: path
          required: true
          schema:
            type: string
            enum:
              - address
              - email
              - phone
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Location'
      responses:
        201:
          description: Create item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
  /admin/locations/{children}/{id}:
    put:
      tags:
        - Location
      summary: update info of Location
      parameters:
        - name: children
          in: path
          description: address | email | phone
          required: true
          schema:
            type: string
            enum:
              - address
              - email
              - phone
        - name: id
          in: path
          description: id of element ('address', 'email', 'phonenumber')
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Location'
      responses:
        201:
          description: Create item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
    delete:
      tags:
        - Location
      summary: Delete item ('address', 'email', 'phone') in Location collection
      description: This can only be done by the logged in user.
      parameters:
        - name: children
          in: path
          description: children is ('address', 'email', 'phone') to return
          required: true
          schema:
            type: string
        - name: id
          in: path
          description: id of element ('address', 'email', 'phone')
          required: true
          schema:
            type: string
      responses:
        201:
          description: Create item success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
...
  /admin/news:
    get:
      tags:
        - New
      summary: Get all posts
      responses:
        201:
          description: Create item success
        422:
          description: Invalid input value
        500:
          description: Internal Server Error
      security:
        - Bearer: []
    post:
      tags:
        - New
      summary: create new post
      description: Create new Sharing
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/New'
      responses:
        201:
          description: Create item success
        422:
          description: Invalid input value
        500:
          description: Internal Server Error
      security:
        - Bearer: []
  /admin/news/{id}:
    get:
      tags:
        - New
      summary: get info post
      description: Returns a single Share
      parameters:
        - name: id
          in: path
          description: ID or Slug of Share to return
          required: true
          schema:
            type: string
      responses:
        200:
          description: 'Listed successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/New'
        404:
          description: New not found.
        500:
          description: Internal Server Error
      security:
        - Bearer: []
    put:
      tags:
        - New
      summary: update new
      description: This can only be done by the logged in user.
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/New'

      responses:
        200:
          description: Delete success.
        404:
          description: Share not found.
        500:
          description: Internal Server Error.
      security:
        - Bearer: []
    delete:
      tags:
        - New
      summary: delete new
      description: Delete Shares
      parameters:
        - name: id
          in: path
          description: ID of User to return
          required: true
          schema:
            type: string
      responses:
        200:
          description: Delete success.
        404:
          description: Share not found.
        500:
          description: Internal Server Error
      security:
        - Bearer: []
...
  /admin/slides:
    get:
      tags:
        - Slide
      summary: get all slides
      responses:
        201:
          description: Create slide success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
    post:
      tags:
        - Slide
      summary: create slide
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Slide'
      responses:
        201:
          description: Create slide success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
  /admin/slides/{id}:
    get:
      tags:
        - Slide
      summary: get info about slide
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        201:
          description: Update slide success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
    put:
      tags:
        - Slide
      summary: update slide by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Slide'
      responses:
        201:
          description: Update slide success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
    delete:
      tags:
        - Slide
      summary: delete slide by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        201:
          description: Update slide success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
  /admin/slides-show:
    post:
      tags:
        - Slideshow
      summary: create new slide show
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Slideshow'
      responses:
        201:
          description: Update slide success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
  /admin/slides-show/{id}:
    get:
      tags:
        - Slideshow
      summary: get info about slides show
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        201:
          description: Update slide success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
    put:
      tags:
        - Slideshow
      summary: update slides show by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Slideshow'
      responses:
        201:
          description: Update slide show success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
    delete:
      tags:
        - Slideshow
      summary: delete slides show by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        201:
          description: Delete slide show success
        404:
          description: Invalid input value
        422:
          description: Create Error
        500:
          description: Internal Server Error
      security:
        - Bearer: []
components:
  schemas:
...
    Location:
      type: object
      properties:
        addressItem:
          type: object
          properties:
            area:
              type: string
              example: HN
              default: TeamPosition
              description: teasefd
            address:
              type: string
              example: 11 Thái Thuận, Phường An Phú, Tp. Thủ Đức, Tp. Hồ Chí Minh
            map:
              type: string
              description: address
        emailItem:
          type: object
          properties:
            title:
              type: string
              example: Blockchain & AI
            email:
              type: string
              example: phongnguyenw@gmail.com
            type:
              type: string
              example: chăm sóc
        phoneItem:
          type: object
          properties:
            title:
              type: string
              example: Blockchain & AI
            phone:
              type: string
              example: 0123456789
            type:
              type: string
              example: chăm sóc
    New:
      type: object
      required:
        - title
        - summary
        - content
        - author
        - mode
        - tags
        - categories
      properties:
        title:
          type: string
          example: Sgod
        summary:
          type: string
          example: Sgod & Blockchain
        content:
          type: string
          example: Sgod CV & Blockchain
        author:
          type: string
          example: Sgod
        mode:
          enum:
            - pending
            - public
            - private
            - hidden
        tags:
          type: array
          items:
            type: string
            example: Sgod
        categories:
          type: array
          items:
            type: string
            example: 633f9d90cc7db2ab77ba503d
        image:
          type: string
          format: binary
        time_public:
          type: string
          example: 2023-04-03T07:37:01.292+00:00
...
    Slide:
      type: object
      required:
        - title
        - description
        - navigate
      properties:
        image:
          type: string
          format: binary
        title:
          type: string
          example: sgod slide
        description:
          type: string
          example: sgod slide description
        navigate:
          type: string
          example: image/slide/sgod-slide.png
    Slideshow:
      type: object
      required:
        - name
        - slide
        - effect
      properties:
        name:
          type: string
          example: sgod show 1
        slides:
          type: array
          items:
            type: string
          example: ['6465998e3e6ab4572157fb63', '6465998e3e6ab4572887ad66']
        effect:
          type: string
          enum:
            - 3D Effect
            - Cube
            - Autoplay
            - Default
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer